+++ internal/repository/user.go
@@ -9,26 +9,26 @@ import (
 
 type UserStorer interface {
 	GetUserByEmail(ctx context.Context, email string) (user User, err error)
-	GetRoleByName(ctx context.Context, name string) (roleId int, err error)
+	GetRoleByName(ctx context.Context, name string) (roleId int64, err error)
 	CreateNewUser(ctx context.Context, user dto.User) (resp User, err error)
 	GetGradeByName(ctx context.Context, name string) (grade Grade, err error)
-	GetRewardMultiplier(ctx context.Context) (value int, err error)
+	GetRewardMultiplier(ctx context.Context) (value int64, err error)
 	SyncData(ctx context.Context, updateData dto.User) (err error)
 }
 
 // User - basic struct representing a User
 type User struct {
-	Id                  int           `db:"id"`
+	Id                  int64         `db:"id"`
 	EmployeeId          string        `db:"employee_id"`
 	FirstName           string        `db:"first_name"`
 	LastName            string        `db:"last_name"`
 	Email               string        `db:"email"`
 	ProfileImageURL     string        `db:"profile_image_url"`
-	GradeId             int           `db:"grade_id"`
+	GradeId             int64         `db:"grade_id"`
 	Designation         string        `db:"designation"`
-	RoleID              int           `db:"role_id"`
-	RewardsQuotaBalance int           `db:"reward_quota_balance"`
-	Status              int           `db:"status"`
+	RoleID              int64         `db:"role_id"`
+	RewardsQuotaBalance int64         `db:"reward_quota_balance"`
+	Status              int64         `db:"status"`
 	SoftDelete          bool          `db:"soft_delete"`
 	SoftDeleteBy        sql.NullInt64 `db:"soft_delete_by"`
 	SoftDeleteOn        sql.NullTime  `db:"soft_delete_on"`
@@ -36,12 +36,12 @@ type User struct {
 }
 
 type Role struct {
-	ID   int    `db:"id"`
+	ID   int64  `db:"id"`
 	Name string `db:"name"`
 }
 
 type Grade struct {
-	Id     int    `db:"id"`
+	Id     int64  `db:"id"`
 	Name   string `db:"name"`
-	Points int    `db:"points"`
+	Points int64  `db:"points"`
 }
